backend/pyproject.toml
[project]
name = "mammothon-backend"
version = "0.1.0"
description = "Backend for mammothon 2025 Q1, Attested Image-Editing Stack"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "Pillow>=10.2.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.1",
    "ImageHash>=4.3.2"
]
